pyproject.toml
[project]
name = "data-energineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
]
